Make terminal rendering work in popout windows
Add support for the parent element (what is passed to Terminal.open)
being in a different (same origin) window. Accesses to DOM APIs such
as requestAnimationFrame and devicePixelRatio thus need to be scoped
to the corrent window, instead of assuming that it's the same as the
window/global scope where the code is running.

This is done by inferring a parent window at the creation time, and
then passing it around to the various services that need it. An
alternative would have been to infer it from the passed in DOM nodes
(via element.ownerDocument.defaultView), but it may not be correct
depending on how those elemnets were created or if they were not
attached to the DOM yet.

This should also be applicable when the parent element is an iframe.